// src/models/thunks.ts
export enum Thunks {
  Register = "user/register",
  Token = "user/token",
  UserByToken = "user/userByToken",
  Users = "user/allUsers",
  UserById = "user/userById",
  Update = "user/update",
  Avatar = "user/updateAvatar",
  CreatePost = "post/createPost",
  MyPosts = "post/fetchMyPosts",
  AllPosts = "post/fetchAllPosts",
  SinglePost = "post/fetchSinglePost",
  DeletePost = "post/deletePost",
  EditPost = "post/editPost",
  LikePost = "post/likePost",
  PostImage = "post/uploadPostImage",
  CreateComment = "comment/createComment",
  Comments = "comment/fetchComments",
  DeleteComment = "comment/deleteComment",
  EditComment = "comment/editComment",
  LikeComment = "comment/likeComment",
}